refactor. Booking, Product 리펙터링 및 엔티티 리펙터링
.

product 컴파일오류 수정

+++ BE/src/main/java/team07/airbnb/data/booking/dto/PriceInfo.java
@@ -8,16 +8,16 @@ import lombok.Getter;
 @Getter
 public class PriceInfo {
 
-    private long roughTotalPrice;
-    private long discountPrice;
-    private long serviceFee;
-    private long accommodationFee;
+    private int roughTotalPrice;
+    private int discountPrice;
+    private int serviceFee;
+    private int accommodationFee;
 
     public boolean isEmpty() {
         return roughTotalPrice == 0 || discountPrice == 0 || serviceFee == 0 || accommodationFee == 0;
     }
 
-    public static PriceInfo of(long roughTotalPrice, long discountPrice, long serviceFee, long accommodationFee) {
+    public static PriceInfo of(int roughTotalPrice, int discountPrice, int serviceFee, int accommodationFee) {
         return new PriceInfo(
                 roughTotalPrice,
                 discountPrice,
