+++ BE/src/main/java/team07/airbnb/data/booking/dto/response/BookingCancelResponse.java
@@ -2,10 +2,10 @@ package team07.airbnb.data.booking.dto.response;
 
 public record BookingCancelResponse(
         Long bookingId,
-        long cancelFee
+        int cancelFee
 ) {
 
-    public static BookingCancelResponse of(Long bookingId, long cancelFee) {
+    public static BookingCancelResponse of(Long bookingId, int cancelFee) {
         return new BookingCancelResponse(bookingId, cancelFee);
     }
 }
