+++ BE/src/main/java/team07/airbnb/data/discount/beans/DiscountPolicy.java
@@ -3,5 +3,5 @@ package team07.airbnb.data.discount.beans;
 
 public interface DiscountPolicy {
 
-    long getDiscountPrice(long price);
+    int getDiscountPrice(int price);
 }
