+++ BE/src/main/java/team07/airbnb/data/discount/beans/WeeklyRateDiscountPolicy.java
@@ -15,7 +15,7 @@ public class WeeklyRateDiscountPolicy implements DiscountPolicy {
     }
 
     @Override
-    public long getDiscountPrice(long price) {
+    public int getDiscountPrice(int price) {
         return RateCalculateUtil.calculate(discountRate, price);
     }
